feat: ✨ improve scroller plugin
