+++ packages/x6-common/src/css-loader/index.ts
@@ -1,0 +1,3 @@
+import * as CssLoader from './loader'
+
+export { CssLoader }
