+++ packages/x6-common/src/dom/main.ts
@@ -7,6 +7,7 @@ export * from './prefix'
 export * from './selection'
 export * from './css'
 export * from './data'
+export * from './prop'
 
 // svg
 // ---
