+++ packages/x6-common/src/index.ts
@@ -21,4 +21,5 @@ export * from './dictionary'
 export * from './registry'
 export * from './modifier'
 export * from './animation'
+export * from './css-loader'
 export * from './types'
