+++ packages/x6-next/src/graph/css.ts
@@ -1,4 +1,4 @@
-import { Platform } from '@antv/x6-common'
+import { CssLoader } from '@antv/x6-common'
 import { Config } from '../config'
 import { content } from '../style/raw'
 import { Base } from './base'
@@ -6,43 +6,12 @@ import { Base } from './base'
 export class CSSManager extends Base {
   protected init() {
     if (Config.autoInsertCSS) {
-      CSSManager.ensure()
+      CssLoader.ensure('core', content)
     }
   }
 
   @CSSManager.dispose()
   dispose() {
-    CSSManager.clean()
-  }
-}
-
-export namespace CSSManager {
-  let styleElement: HTMLStyleElement | null
-  let counter = 0
-
-  export function ensure() {
-    counter += 1
-    if (counter > 1) return
-
-    if (!Platform.isApplyingHMR()) {
-      styleElement = document.createElement('style')
-      styleElement.setAttribute('type', 'text/css')
-      styleElement.textContent = content
-
-      const head = document.querySelector('head') as HTMLHeadElement
-      if (head) {
-        head.insertBefore(styleElement, head.firstChild)
-      }
-    }
-  }
-
-  export function clean() {
-    counter -= 1
-    if (counter > 0) return
-
-    if (styleElement && styleElement.parentNode) {
-      styleElement.parentNode.removeChild(styleElement)
-    }
-    styleElement = null
+    CssLoader.clean('core')
   }
 }
